chore: move storybook to top-level workspace
