+++ .storybook/main.js
@@ -2,7 +2,7 @@
  * @type {import('@storybook/core-common').StorybookConfig}
  */
 module.exports = {
-  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
+  stories: ['../packages/**/src/**/*.stories.mdx', '../packages/**/src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
     {name: '@storybook/addon-essentials', options: {backgrounds: false}},
     '@storybook/addon-storysource',
