+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "primer",
+  "version": "0.0.0",
   "private": true,
   "workspaces": [
     "packages/*",
@@ -18,6 +19,10 @@
     "lint:md": "markdownlint-cli2 \"**/*.{md,mdx}\" \"!.github\" \"!.changeset\" \"!**/node_modules/**\" \"!**/CHANGELOG.md\"",
     "prepare": "husky install",
     "setup": "./script/setup",
+    "start": "concurrently npm:start:*",
+    "start:docs": "cd docs && npm run develop",
+    "start:storybook": "STORYBOOK=true start-storybook -p 6006",
+    "build:storybook": "script/build-storybook",
     "test": "jest",
     "test:coverage": "npm run test -- --coverage",
     "test:update": "npm run test -- --updateSnapshot",
@@ -30,6 +35,18 @@
     "@babel/preset-env": "7.21.5",
     "@github/markdownlint-github": "^0.3.0",
     "@github/prettier-config": "0.0.6",
+    "@storybook/addon-a11y": "6.5.16",
+    "@storybook/addon-actions": "6.5.16",
+    "@storybook/addon-essentials": "6.5.16",
+    "@storybook/addon-interactions": "6.5.14",
+    "@storybook/addon-links": "6.5.16",
+    "@storybook/addon-storysource": "6.5.14",
+    "@storybook/builder-webpack5": "6.5.16",
+    "@storybook/manager-webpack5": "6.5.14",
+    "@storybook/react": "6.5.16",
+    "@storybook/test-runner": "0.10.0",
+    "@storybook/testing-library": "0.0.13",
+    "@storybook/theming": "6.5.16",
     "@types/jest": "29.4.0",
     "@types/jest-axe": "3.5.5",
     "@typescript-eslint/eslint-plugin": "5.54.1",
@@ -60,6 +77,11 @@
     "markdownlint-cli2": "^0.7.1",
     "markdownlint-cli2-formatter-pretty": "0.0.3",
     "prettier": "2.8.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-test-renderer": "18.2.0",
+    "storybook-addon-turbo-build": "1.1.0",
+    "storybook-css-modules": "1.0.8",
     "typescript": "4.9.5"
   },
   "prettier": "@github/prettier-config"
