+++ packages/react/package.json
@@ -36,9 +36,6 @@
     "clean": "rimraf dist lib lib-esm css",
     "start": "concurrently npm:start:*",
     "start:docs": "cd docs && npm run develop",
-    "start:storybook": "STORYBOOK=true start-storybook -p 6006",
-    "build:storybook": "script/build-storybook",
-    "build:storybook:visual-testing": "build-storybook",
     "build:components.json": "ts-node script/components-json/build",
     "release": "npm run build && changeset publish",
     "size": "size-limit"
@@ -119,18 +116,6 @@
     "@rollup/plugin-typescript": "11.1.0",
     "@rollup/plugin-virtual": "3.0.1",
     "@size-limit/preset-big-lib": "8.2.4",
-    "@storybook/addon-a11y": "6.5.16",
-    "@storybook/addon-actions": "6.5.16",
-    "@storybook/addon-essentials": "6.5.16",
-    "@storybook/addon-interactions": "6.5.14",
-    "@storybook/addon-links": "6.5.16",
-    "@storybook/addon-storysource": "6.5.14",
-    "@storybook/builder-webpack5": "6.5.16",
-    "@storybook/manager-webpack5": "6.5.14",
-    "@storybook/react": "6.5.16",
-    "@storybook/test-runner": "0.10.0",
-    "@storybook/testing-library": "0.0.13",
-    "@storybook/theming": "6.5.16",
     "@testing-library/dom": "9.2.0",
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "14.0.0",
@@ -189,8 +174,6 @@
     "rollup-plugin-visualizer": "5.9.0",
     "semver": "7.5.0",
     "size-limit": "8.2.4",
-    "storybook-addon-turbo-build": "1.1.0",
-    "storybook-css-modules": "1.0.8",
     "styled-components": "4.4.1",
     "terser": "5.16.1",
     "ts-node": "10.9.1",
